requirements.txt
aiohttp
aiosignal
alembic
annotated-types
anyio
asttokens
attrs
beautifulsoup4
bs4
cachetools
certifi
charset-normalizer
click
comm
contourpy
coverage
cycler
dataclasses-json
debugpy
decorator
Deprecated
dirtyjson
distro
et-xmlfile
executing
fastapi
fastjsonschema
fonttools
frozenlist
fsspec
fuzzywuzzy
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud
google-cloud-core
google-cloud-pubsub
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
groq
grpc-google-iam-v1
grpcio
grpcio-status
gspread
gspread-dataframe
h11
httpcore
httplib2
httpx
idna
iniconfig
ipykernel
ipython
jedi
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
kiwisolver
langchain
langchain-community
langchain-core
langchain-groq
langchain-openai
langchain-text-splitters
langsmith
Levenshtein
llama-index
llama-index-agent-openai
llama-index-cli
llama-index-core
llama-index-embeddings-openai
llama-index-indices-managed-llama-cloud
llama-index-legacy
llama-index-llms-openai
llama-index-multi-modal-llms-openai
llama-index-program-openai
llama-index-question-gen-openai
llama-index-readers-file
llama-index-readers-llama-parse
llama-parse
llamaindex-py-client
Mako
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
multidict
mypy-extensions
nbformat
nest-asyncio
networkx
nltk
numpy
oauthlib
openai
openpyxl
orjson
packaging
pandas
parso
pexpect
pillow
pip
platformdirs
plotly
pluggy
prompt-toolkit
proto-plus
protobuf
psutil
ptyprocess
pure-eval
pyasn1
pyasn1-modules
pydantic
pydantic_core
Pygments
PyMuPDF
PyMuPDFb
PyMySQL
pyparsing
pypdf
pytest
pytest-cov
pytest-mock
python-dateutil
python-dotenv
python-Levenshtein
pytz
PyYAML
pyzmq
rapidfuzz
referencing
regex
requests
requests-oauthlib
rpds-py
rsa
scipy
seaborn
setuptools
six
sniffio
soupsieve
SQLAlchemy
stack-data
starlette
StrEnum
striprtf
tenacity
tiktoken
tornado
tqdm
traitlets
typing_extensions
typing-inspect
tzdata
uritemplate
urllib3
uvicorn
wcwidth
wrapt
yarl
